research-agent/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-dotenv>=1.0.0
dateparser>=1.1.0
streamlit>=1.28.0
anthropic>=0.40.0
